pyproject.toml
[project]
name = "technical-doc-generator"
version = "0.1.0"
description = "AI-powered Technical Document Generator using Context7 MCP and AWS Bedrock"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI and web server
    "fastapi>=0.121.3",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    # AWS SDK and LangChain
    "boto3>=1.41.2",
    "langchain-aws>=1.0.0",
    "langchain-openai>=1.0.3",
    # HTTP client for MCP
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    # Template and markdown processing
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "pymdown-extensions>=10.0",
    # Environment and config
    "python-dotenv>=1.2.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Utilities
    "colorama>=0.4.6",
    "requests>=2.32.5",
    "pyyaml>=6.0.3",
    # Development tools
    "pre-commit>=4.3.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "botocore[crt]>=1.41.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "jupyterlab>=4.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
